+++ examples/databases/normal-load/install.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-set -eux
+set -eu
+set -o pipefail
+trap "trap - SIGTERM && kill -- -$$" SIGINT SIGTERM EXIT
+
 
 export NAMESPACE=normal-load
 export SCENARIO=$(dirname -- "$0")/manifest.yml
@@ -11,13 +14,13 @@ export DASHBOARDS=(summary ingleton ycsb)
 # Prepare the Reporting folder
 mkdir -p "${REPORTS}"
 
-# Submit the scenario and follow logs
-kubectl-frisbee submit test "${NAMESPACE}" "${SCENARIO}" "${DEPENDENCIES[@]}"
-
 # Copy the manifest
 cp "${SCENARIO}" "${REPORTS}"
 
-# wait for the scenario to be submitted
+# Submit the scenario and follow logs
+kubectl-frisbee submit test "${NAMESPACE}" "${SCENARIO}" "${DEPENDENCIES[@]}"
+
+# Give a headstart
 sleep 10
 
 # Report the scenario