+++ examples/databases/sstable-bitrot/install.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-set -eux
+set -eu
+set -o pipefail
+trap "trap - SIGTERM && kill -- -$$" SIGINT SIGTERM EXIT
 
 export NAMESPACE=sstable-bitrot
 export SCENARIO=$(dirname -- "$0")/manifest.yml
@@ -10,14 +12,14 @@ export DEPENDENCIES=(./charts/system/ ./charts/databases/cockroachdb ./charts/da
 # Prepare the Reporting folder
 mkdir -p "${REPORTS}"
 
-# Submit the scenario and follow logs
-kubectl-frisbee submit test "${NAMESPACE}" "${SCENARIO}" "${DEPENDENCIES[@]}"
-
 # Copy the manifest
 cp "${SCENARIO}" "${REPORTS}"
 
-# wait for the scenario to be submitted
+# Submit the scenario and follow the failing client's logs
+kubectl-frisbee submit test "${NAMESPACE}" "${SCENARIO}" "${DEPENDENCIES[@]}"
+
+# Give a headstart
 sleep 10
 
-# Report the scenario
+# Download test report
 kubectl-frisbee report test "${NAMESPACE}" "${REPORTS}" --pdf --data --aggregated-pdf --wait