+++ examples/federated_learning/ml-backend/install.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+set -eu
+set -o pipefail
+trap "trap - SIGTERM && kill -- -$$" SIGINT SIGTERM EXIT
+
 export NAMESPACE=ml-backend
 export SCENARIO=$(dirname -- "$0")/manifest.yml
 export REPORTS=${HOME}/frisbee-reports/${NAMESPACE}/
@@ -8,13 +12,13 @@ export DEPENDENCIES=(./charts/system/ ./charts/federated-learning/fedbed/)
 # Prepare the Reporting folder
 mkdir -p "${REPORTS}"
 
-# Submit the scenario and follow logs
-kubectl-frisbee submit test "${NAMESPACE}" "${SCENARIO}" "${DEPENDENCIES[@]}"
-
 # Copy the manifest
 cp "${SCENARIO}" "${REPORTS}"
 
-# wait for the scenario to be submitted
+# Submit the scenario and follow server logs
+kubectl-frisbee submit test "${NAMESPACE}" "${SCENARIO}" "${DEPENDENCIES[@]}" --logs server |& tee  "${REPORTS}"/logs &
+
+# Give a headstart
 sleep 10
 
 # Report the scenario